+++ src/Component/Testing/src/Functional/BaseTestCase.php
@@ -11,10 +11,12 @@
 
 namespace Symfony\Cmf\Component\Testing\Functional;
 
-use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
+use Doctrine\Bundle\PHPCRBundle\Test\RepositoryManager;
 use Symfony\Bundle\FrameworkBundle\Client;
-use Symfony\Component\HttpFoundation\Response;
+use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
+use Symfony\Cmf\Component\Testing\Functional\DbManager\PhpcrDecorator;
 use Symfony\Component\DependencyInjection\Container;
+use Symfony\Component\HttpFoundation\Response;
 
 /**
  * The base class for Functional and Web tests.
@@ -29,7 +31,7 @@ abstract class BaseTestCase extends WebTestCase
      *
      * @var array
      */
-    protected $dbManagers = array();
+    protected $dbManagers = [];
 
     /**
      * @var Container
@@ -53,7 +55,7 @@ abstract class BaseTestCase extends WebTestCase
      */
     protected function getKernelConfiguration()
     {
-        return array();
+        return [];
     }
 
     /**
@@ -112,6 +114,10 @@ abstract class BaseTestCase extends WebTestCase
             $type
         );
 
+        if ('phpcr' === strtolower($type) && class_exists(RepositoryManager::class)) {
+            $className = PhpcrDecorator::class;
+        }
+
         if (!class_exists($className)) {
             throw new \InvalidArgumentException(sprintf(
                 'Test DBManager "%s" does not exist.',
@@ -126,12 +132,26 @@ abstract class BaseTestCase extends WebTestCase
         return $dbManager;
     }
 
+    public static function getKernelClass()
+    {
+        if (isset($_SERVER['KERNEL_CLASS']) || isset($_ENV['KERNEL_CLASS'])) {
+            $class = isset($_SERVER['KERNEL_CLASS']) ? $_SERVER['KERNEL_CLASS'] : $_ENV['KERNEL_CLASS'];
+            if (!class_exists($class)) {
+                throw new \RuntimeException(sprintf('Class "%s" doesn\'t exist or cannot be autoloaded. Check that the KERNEL_CLASS value in phpunit.xml matches the fully-qualified class name of your Kernel or override the %s::createKernel() method.', $class, static::class));
+            }
+
+            return $class;
+        }
+
+        return parent::getKernelClass();
+    }
+
     /**
      * {@inheritdoc}
      *
      * This is overriden to set the default environment to 'phpcr'
      */
-    protected static function createKernel(array $options = array())
+    protected static function createKernel(array $options = [])
     {
         // default environment is 'phpcr'
         if (!isset($options['environment'])) {