+++ src/Component/Testing/src/Phpunit/DatabaseTestListener.php
@@ -11,31 +11,40 @@
 
 namespace Symfony\Cmf\Component\Testing\Phpunit;
 
-use Symfony\Component\Process\ProcessBuilder;
-use Symfony\Component\Process\PhpExecutableFinder;
 use Doctrine\Common\DataFixtures\Purger;
+use Symfony\Component\Process\PhpExecutableFinder;
+use Symfony\Component\Process\Process;
 
 class DatabaseTestListener implements \PHPUnit_Framework_TestListener
 {
     protected static $currentSuite;
-    private $processBuilder;
-    private $prefix = array();
 
-    public function __construct($processBuilder = null)
+    private $processCallable;
+
+    private $prefix = [];
+
+    public function __construct(callable $processCallable = null)
     {
-        if (null === $processBuilder) {
-            $this->processBuilder = new ProcessBuilder();
-            $phpExecutableFinder = new PhpExecutableFinder();
-            $phpExecutable = $phpExecutableFinder->find(false);
-            if (false === $phpExecutable) {
-                throw new \RuntimeException('No PHP executable found on the current system.');
-            }
-
-            // Symfony 2.3 does not support array prefix, so we have to implement it ourselves
-            $this->prefix = array($phpExecutable, __DIR__.'/../../bin/console');
-        } else {
-            $this->processBuilder = $processBuilder;
+        $phpExecutableFinder = new PhpExecutableFinder();
+        $phpExecutable = $phpExecutableFinder->find(false);
+        if (false === $phpExecutable) {
+            throw new \RuntimeException('No PHP executable found on the current system.');
         }
+
+        // Symfony 2.3 does not support array prefix, so we have to implement it ourselves
+        $this->prefix = [$phpExecutable, __DIR__.'/../../bin/console'];
+        $this->processCallable = $processCallable;
+    }
+
+    public function getProcess($arguments)
+    {
+        if (is_callable($this->processCallable)) {
+            $callable = $this->processCallable;
+
+            return $callable($arguments);
+        }
+
+        return new Process($arguments);
     }
 
     public function addError(\PHPUnit_Framework_Test $test, \Exception $e, $time)
@@ -68,11 +77,13 @@ class DatabaseTestListener implements \PHPUnit_Framework_TestListener
             case 'orm':
                 $db = $test->getDbManager('ORM');
                 $purger = new Purger\ORMPurger($db->getOm());
+
                 break;
 
             case 'phpcr':
                 $db = $test->getDbManager('PHPCR');
                 $purger = new Purger\PHPCRPurger($db->getOm());
+
                 break;
 
             default:
@@ -94,10 +105,12 @@ class DatabaseTestListener implements \PHPUnit_Framework_TestListener
         switch ($suite->getName()) {
             case 'orm':
                 $this->setUpOrmDatabase($suite);
+
                 break;
 
             case 'phpcr':
                 $this->setUpPhpcrDatabase($suite);
+
                 break;
 
             default:
@@ -112,17 +125,13 @@ class DatabaseTestListener implements \PHPUnit_Framework_TestListener
         echo PHP_EOL.PHP_EOL;
 
         // initialize PHPCR DBAL (new way)
-        $process = $this->processBuilder
-            ->setArguments(array_merge($this->prefix, array('doctrine:phpcr:init:dbal', '--drop', '--force')))
-            ->getProcess();
+        $process = $this->getProcess(array_merge($this->prefix, ['doctrine:phpcr:init:dbal', '--drop', '--force']));
 
         $process->run();
 
         if (!$process->isSuccessful()) {
             // try initializing the old way (Jackalope <1.2)
-            $process = $this->processBuilder
-                ->setArguments(array_merge($this->prefix, array('doctrine:phpcr:init:dbal', '--drop')))
-                ->getProcess();
+            $process = $this->getProcess(array_merge($this->prefix, ['doctrine:phpcr:init:dbal', '--drop']));
 
             $process->run();
 
@@ -135,9 +144,7 @@ class DatabaseTestListener implements \PHPUnit_Framework_TestListener
         }
 
         // initialize repositories
-        $process = $this->processBuilder
-            ->setArguments(array_merge($this->prefix, array('doctrine:phpcr:repository:init')))
-            ->getProcess();
+        $process = $this->getProcess(array_merge($this->prefix, ['doctrine:phpcr:repository:init']));
 
         $process->run();
 
@@ -155,9 +162,7 @@ class DatabaseTestListener implements \PHPUnit_Framework_TestListener
     {
         echo PHP_EOL.PHP_EOL;
 
-        $process = $this->processBuilder
-            ->setArguments(array_merge($this->prefix, array('doctrine:schema:drop', '--env=orm', '--force')))
-            ->getProcess();
+        $process = $this->getProcess(array_merge($this->prefix, ['doctrine:schema:drop', '--env=orm', '--force']));
 
         $process->run();
 
@@ -168,9 +173,7 @@ class DatabaseTestListener implements \PHPUnit_Framework_TestListener
             return;
         }
 
-        $process = $this->processBuilder
-            ->setArguments(array_merge($this->prefix, array('doctrine:database:create', '--env=orm')))
-            ->getProcess();
+        $process = $this->getProcess(array_merge($this->prefix, ['doctrine:database:create', '--env=orm']));
 
         $process->run();
 
@@ -181,9 +184,7 @@ class DatabaseTestListener implements \PHPUnit_Framework_TestListener
             return;
         }
 
-        $process = $this->processBuilder
-            ->setArguments(array_merge($this->prefix, array('doctrine:schema:create', '--env=orm')))
-            ->getProcess();
+        $process = $this->getProcess(array_merge($this->prefix, ['doctrine:schema:create', '--env=orm']));
         $process->run();
 
         if (!$process->isSuccessful()) {
@@ -198,13 +199,11 @@ class DatabaseTestListener implements \PHPUnit_Framework_TestListener
 
     public function endTestSuite(\PHPUnit_Framework_TestSuite $suite)
     {
-        if (!in_array($suite->getName(), array('phpcr', 'orm'))) {
+        if (!in_array($suite->getName(), ['phpcr', 'orm'])) {
             return;
         }
 
-        $process = $this->processBuilder
-            ->setArguments(array_merge($this->prefix, array('doctrine:database:drop', '--force')))
-            ->getProcess();
+        $process = $this->getProcess(array_merge($this->prefix, ['doctrine:database:drop', '--force']));
 
         $process->run();