+++ src/Component/Testing/tests/HttpKernel/TestKernelTest.php
@@ -11,21 +11,26 @@
 
 namespace Symfony\Cmf\Component\Testing\Tests\HttpKernel;
 
+use Doctrine\Bundle\DoctrineBundle\DoctrineBundle;
+use Doctrine\Bundle\PHPCRBundle\DoctrinePHPCRBundle;
 use Symfony\Bundle\FrameworkBundle\FrameworkBundle;
 use Symfony\Bundle\SecurityBundle\SecurityBundle;
 use Symfony\Bundle\TwigBundle\TwigBundle;
-use Doctrine\Bundle\DoctrineBundle\DoctrineBundle;
-use Doctrine\Bundle\PHPCRBundle\DoctrinePHPCRBundle;
+use Symfony\Component\HttpKernel\Kernel;
 
 class TestKernelTest extends \PHPUnit_Framework_TestCase
 {
+    /**
+     * @var Kernel
+     */
     private $kernel;
+
     private $mockBundle;
 
     protected function setUp()
     {
         $this->kernel = $this->getMockBuilder('Symfony\Cmf\Component\Testing\HttpKernel\TestKernel')
-            ->setConstructorArgs(array('test', true))
+            ->setConstructorArgs(['test', true])
             ->getMockForAbstractClass();
 
         $this->mockBundle = $this->createMock('Symfony\Component\HttpKernel\Bundle\BundleInterface');