+++ src/Component/Testing/CHANGELOG.md
@@ -1,6 +1,20 @@
 Changelog
 =========
 
+2.1.0 (unreleased)
+------------------
+
+* **2017-01-18**: Introduce a `TestCompilePass` to make services public, which are needed 
+in i.e. a WebTest.
+ * **2017-11-08**: Removed php 5.6 and 7.0 support, removed Symfony 3.0.* and 3.1.* support
+ introduce KERNEL_CLASS handling to avoid deprecated KERNEL_DIR, removed usage of `ProcessBuilder`
+
+2.0.1
+-----
+
+ * **2018-01-02**: Remove dependency on symfony/symfony in favor of more specific packages. Its a
+   bug of the package using testing if it does not declare its dependencies.
+
 2.0.0
 -----